Move vocprez queries to under queries directory
